Simplified and streamlined training facility.
We now have a single train function which can configured solely by passing it a
Namespace or NamedTuple, etc. with all the necessary parameters.
It also brings a bunch of improved(streamlined and accelerated) code structures
with it and reduced unnecessary bulk code and modularisation in with the
monitoring facility (this has now been integrated into a single wrapper, which
reduces flexibility but greatly increases clarity of code).